Update module settings with new config
