+++ README.md
@@ -19,20 +19,20 @@ Run prediction models on [Digital Earth Australia](https://www.ga.gov.au/dea) Ne
 ### Build/run locally
 1. Install dependencies in virtual enviroment: `pipenv install`
 1. Initialise virtual environment: `pipenv shell`
-1. Run the NRT predict process: `python -m nrt_predict s3://dea-public-data/L2/sentinel-2-nrt/S2MSIARD/2021-03-25/S2A_OPER_MSI_ARD_TL_VGS4_20210325T014951_A030057_T56HKH_N02.09`
+1. Run the NRT predict process: `python -m nrtpredict s3://dea-public-data/L2/sentinel-2-nrt/S2MSIARD/2021-03-25/S2A_OPER_MSI_ARD_TL_VGS4_20210325T014951_A030057_T56HKH_N02.09`
 
 ### Alternative (if you don't want to use pipenv to manage your virutal environment)
 1. Generate regular *requirements.txt*: `pipenv run pip freeze > requirements.txt`
 1. (Optional) activate your environment (eg. `source ./venv/bin/activate` etc).
 1. Install dependencies via pip: `pip install -r requirements.txt`
-1. Run the NRT predict process: `python -m nrt_predict s3://dea-public-data/L2/sentinel-2-nrt/S2MSIARD/2021-03-25/S2A_OPER_MSI_ARD_TL_VGS4_20210325T014951_A030057_T56HKH_N02.09`
+1. Run the NRT predict process: `python -m nrtpredict s3://dea-public-data/L2/sentinel-2-nrt/S2MSIARD/2021-03-25/S2A_OPER_MSI_ARD_TL_VGS4_20210325T014951_A030057_T56HKH_N02.09`
 
 ### Build/Run the Docker container
 1. `./run_in_docker.sh`
 
 ### Run a model in Docker
 1. `docker exec -it nft-predict bash`
-1. `python -m nrt_predict s3://dea-public-data/L2/sentinel-2-nrt/S2MSIARD/2021-03-25/S2A_OPER_MSI_ARD_TL_VGS4_20210325T014951_A030057_T56HKH_N02.09`
+1. `python -m nrtpredict s3://dea-public-data/L2/sentinel-2-nrt/S2MSIARD/2021-03-25/S2A_OPER_MSI_ARD_TL_VGS4_20210325T014951_A030057_T56HKH_N02.09`
 
 ## Information
