+++ nrtpredict/__main__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+
+from .cli import *
+
+if __name__ == "__main__":
+    cli.cli_entry()
