+++ nrtpredict/models/base.py
