+++ nrtpredict/models/burnscar.py
