+++ nrtpredict/models/hotspots.py
