+++ nrtpredict/tests/test_cli.py
